move styles and helper function on DataSourceSidebar
